Fix for latest versions of Tensorflow and 10.1 CUDA Compatibility.
Hello, nice project you have here. The GazeGan.py file won't run due to a bug with batches. The current code only allows one batch. Anymore than one batch and it won't run. You can test this yourself, and this little patch fixes it up to the latest versions of Tensorflow that are compatible with Python 2.7. Adding and deleting these lines fixes it.